+++ trunk/src/game/ObjectAccessor.h
@@ -53,7 +53,7 @@ class MANGOS_DLL_DECL ObjectAccessor : public MaNGOS::Singleton<ObjectAccessor, 
 
         Creature* GetCreature(Unit const &, uint64);
         Corpse*   GetCorpse(Unit const &u, uint64 guid);
-        Corpse*   GetCorpse(uint32 x, uint32 y, uint32 mapid, uint64 guid);
+        Corpse*   GetCorpse(float x, float y, uint32 mapid, uint64 guid);
         Unit* GetUnit(Unit const &, uint64);
         Player* GetPlayer(Unit const &, uint64);
         GameObject* GetGameObject(Unit const &, uint64);
